Move Participant, Transport node.js specifics to msgflo-nodejs
https://github.com/msgflo/msgflo-nodejs

msgflo now contains only the general tooling/setup code that
can be used with any MsgFlo participant. This makes it clear
that we aim to be a proper polyglot solution, node.js is not special

msgflo now depends on msgflo-nodejs for the transport layer,
but this is just an implementation detail.